requirements.txt
Flask==2.1.2
requests
psycopg2-binary